requirements-travis.txt
attrs>=17,<18
codecov
coverage
flake8==3.4.1
flake8-docstrings==1.1.0
